Patch Level 23a - Mochimo Transitions to Open Beta
Initializing GitHub Public Repository as the Mochimo network transitions to the open beta testing phase, following completion of the closed beta.  Please join the slack channel linked from mochimo.org to discuss the code with the community as well to interact directly with the developers.  Additionally you may use the below link:

https://join.slack.com/t/mochimo/shared_invite/enQtMzA4NDc3NDU3MzY2LTUwOGExMmM1YWFmMzhiYzdmOWQ5MzcxYzE1MmM3YTA5ZGVlNzZjMGJkZTc1YjM3NGQ3MTVlNWMxYzhiYWQ4NDg