CUDA mGPU Utilization Fix
This hotfix addresses the under-utilization of mGPU mining rigs running
the Mochimo Node v2 software. A 10% gain in Haiku/s was also achieved.

To achieve the fix, I implemented a process where each GPU is given a
different seed, and unbound the host loop from cudaDeviceSynchronize()
such that as soon as the GPU is finished processing it's seed, it is
immeditately given another one that was prepared in advance. An
unintended side effect of this implementation is that even a single
GPU will benefit from a small Haikurate (H/s) gain of nearly 10%.
The miner.c file is also formatted/cleaned in preparation of the
ability to compile your choice of miner CUDA/AMD/CPU/NONE in the near
future.