Fix *** Balance Check Failed
This hotfix addresses the  error encountered
while the wallet tries to query a node on the network for the balancce
of an address.

To achieve the fix, I changed the  variable to represent the port
used on MainNet (2095) and replaced the TestNet IP's in
with localhost only. The wallet will then read a  file
containing IP's and fill  with them. As an additional
measure, if the connection attempt to a node returns an error, the next
connection attempt will use the next IP in .