mThread Initialization rev2 (neogen backup)
Revision 2 implements a backup solution for neo-genesis block. These
blocks will eventually be quite weighty and downloading them will take
some time, so upon resync the node will search to see if it has the
appropriate neogen in storage and start with that file instead of
downloading it. Upon failure of the backup neo-genesis block, the node
will resync without it and download a fresh neo-genesis block.

Private (local) test results for a resync are down to less than 2
minutes with a quorum of 8. A fresh initialization of a node still
completes in less than 5 minutes. (these are average results)