Allowing OP_BALANCE Queries to Omit Tag
The new handler function will take a Zero'd tag in a OP_BALANCE request, find an address in the ledger that matches everything except the tag, and return both the address it found and what the balance is.  Useful for wallet account recovery where the seed for the wallet address-chain is known but the current address and/or its tag is not know.