Added v2.3 Node Filter After v2.4 Trigger Block
After the v2.4 trigger block any node running the old code will not be able to speak to the network.  This test is removed some time after the current code fork to enable forward compatibility with the next code version, and then re-added the next time a soft fork occurs.