Pushblock reconfiguration fix
Allowed headless miners more accessibility by reconfiguring the
Pushblock routines. Previously, if a node was also mining, the
candidate block was consumed and converted into miner.tmp.
Similiarly, if it was not mining, miner.tmp did not exist and
the node was unable to provide a candidate block. This will be
finalized in the major miner rewrite commit.