introduce selective __constant__ qualifier
It seems 30 Series Cuda devices, specifically sm_86, performs better
without the __constant__ qualifier on static memory in hashing algos.
So a new definition was introduced to remove the qualifier when
compiling for sm_86 hardware.

Note: the architecture is required to be specified at compile time
for effective use of selective __constant__ qualifier