initial variable PDU protocol implementation
The proposed changes for variable PDU communication between nodes
involves a 3 and 2 stage process for recv() and send(), respectively.

To recv(), we read the PDU header to determine a NODEs capability and
provided buffer length (stage 1), then we recv() the buffer data into
the buffer of a TX (stage 2), and finally, we recv() the trailer data
into the crc16 and EOT of a TX (stage 3).

To send(), we write the PDU header+buffer data (stage 1), followed by
the crc16 and EOT constant (stage 2).

To maintain initial compatibility with PVERSION 4 network operations,
all OP_HELLO operations will be fixed length for this implementation.
This compatibility will become redundant in the switch to PVERSION 5.