add txqe_val() and update tx_nonce/tx_id on clean
tx_val() validates transaction data from a TXQENTRY object, as if
received directly from a wallet or via a node mirror, where the nonce
and id are not yet required. txqe_val(), in comparison, validates the
entire transaction entry, as if read from the chain, where the nonce
and id are required. Additionally, we update the nonce and id after a
clean, ready to be included into the next block.