REVERT THIS COMMIT AFTER wots-improvements FIX
This commit is specifically designed to replace the wots-improvements
branch to it's original state. This is because the improvements are
affecting the operability of validation procedures. These improvements
are low priority and so a FIX for the invalid operation of sig->pk
is on hold.

Once a FIX is provided, it should be committed to the wots-improvments
branch and this commit should be reverted via 'git revert', before
merging the fixes back into the development branch. Ideally, the test
case developed for wots routines will also need updating.